(PC-8207) test the call to notifyWebappLinkSent
